+++ cypress/e2e/navigation.cy.ts
@@ -1,76 +1,8 @@
 // @ts-ignore
-import {core, KalturaPlayer} from '@playkit-js/kaltura-player-js';
+import {core} from '@playkit-js/kaltura-player-js';
+import {mockKalturaBe, loadPlayer, MANIFEST, MANIFEST_SAFARI} from './env';
 
 const {FakeEvent} = core;
-const MANIFEST = `#EXTM3U
-#EXT-X-MEDIA:TYPE=AUDIO,GROUP-ID="audio",LANGUAGE="en",NAME="English",AUTOSELECT=YES,DEFAULT=YES,URI="${location.origin}/media/index_1.m3u8",SUBTITLES="subs"
-#EXT-X-STREAM-INF:PROGRAM-ID=1,BANDWIDTH=509496,RESOLUTION=480x272,AUDIO="audio",SUBTITLES="subs"
-${location.origin}/media/index.m3u8`;
-
-const MANIFEST_SAFARI = `#EXTM3U
-#EXT-X-MEDIA:TYPE=SUBTITLES,GROUP-ID="subs",NAME="English",DEFAULT=NO,AUTOSELECT=YES,FORCED=NO,LANGUAGE="en",URI="${location.origin}/media/index_1.m3u8"
-#EXT-X-STREAM-INF:PROGRAM-ID=1,BANDWIDTH=504265,RESOLUTION=480x272,AUDIO="audio",SUBTITLES="subs"
-${location.origin}/media/index.m3u8`;
-
-const getPlayer = () => {
-  // @ts-ignore
-  return cy.window().then($win => $win.KalturaPlayer.getPlayers()['player-placeholder']);
-};
-
-const preparePage = (navigationConf: Object, playbackConf: Object) => {
-  cy.visit('index.html');
-  return cy.window().then(win => {
-    try {
-      // @ts-ignore
-      var kalturaPlayer = win.KalturaPlayer.setup({
-        targetId: 'player-placeholder',
-        provider: {
-          partnerId: -1,
-          env: {
-            cdnUrl: 'http://mock-cdn',
-            serviceUrl: 'http://mock-api'
-          }
-        },
-        plugins: {
-          navigation: navigationConf,
-          uiManagers: {},
-          kalturaCuepoints: {}
-        },
-        playback: {muted: true, autoplay: true, ...playbackConf}
-      });
-      return kalturaPlayer.loadMedia({entryId: '0_wifqaipd'});
-    } catch (e: any) {
-      return Promise.reject(e.message);
-    }
-  });
-};
-
-const loadPlayer = (navigationConf = {}, playbackConf = {}) => {
-  return preparePage(navigationConf, playbackConf).then(() => getPlayer().then(kalturaPlayer => kalturaPlayer.ready().then(() => kalturaPlayer)));
-};
-
-const checkRequest = (reqBody: any, service: string, action: string) => {
-  return reqBody?.service === service && reqBody?.action === action;
-};
-
-const mockKalturaBe = (entryFixture = 'vod-entry.json', dataFixture = 'navigation-data.json', captionsFixture = 'captions-en-response.json') => {
-  cy.intercept('http://mock-api/service/multirequest', req => {
-    if (checkRequest(req.body[2], 'baseEntry', 'list')) {
-      return req.reply({fixture: entryFixture});
-    }
-    if (checkRequest(req.body[2], 'cuepoint_cuepoint', 'list')) {
-      return req.reply({fixture: dataFixture});
-    }
-    if (checkRequest(req.body[2], 'caption_captionasset', 'serveAsJson')) {
-      return req.reply({fixture: captionsFixture});
-    }
-    if (checkRequest(req.body[2], 'thumbAsset', 'getUrl')) {
-      return req.reply({fixture: 'thumb-url.json'});
-    }
-  });
-  cy.intercept('GET', '**/ks/123', {fixture: 'thumb-asset.jpeg'}).as('getSlides');
-  cy.intercept('GET', '**/vid_sec/*', {fixture: 'thumb-asset.jpeg'}).as('getChapters');
-};
 
 describe('Navigation plugin', () => {
   beforeEach(() => {
