+++ package.json
@@ -12,7 +12,7 @@
     "url": "git+https://github.com/kaltura/playkit-js-navigation.git"
   },
   "dependencies": {
-    "@playkit-js/common": "^1.2.5",
+    "@playkit-js/common": "^1.2.9",
     "@playkit-js/playkit-js-ui": "^0.77.1",
     "@playkit-js/ui-managers": "^1.3.8"
   },
